deezer/view_sets/authenticated_generic_view_set.py
from rest_framework.permissions import IsAuthenticated
from rest_framework.viewsets import GenericViewSet


class AuthenticatedGenericViewSet(GenericViewSet):
    permission_classes = [IsAuthenticated]
