requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-guardian==2.3.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
requests==2.25.1
django-utils-six
django_facebook
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0